requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
openrouteservice==2.3.3
pandas==2.2.1
numpy==1.26.4
pydantic==2.6.4
python-dotenv==1.0.1
openai==1.23.2
requests==2.31.0
python-multipart==0.0.9
firebase-admin==6.4.0
